requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
pillow
transformers
sentence-transformers
motor
python-dotenv
boto3
pytest
torch
pandas